requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
playwright
aiohttp
apscheduler
pyyaml